Create BoolMaskBySizes
The scrypt takes two arguments - a list with family instances, representing structural opening families, and a list with lists of elements of the same type. It identifies instances from the second list, located at the same X and Y coordinates and having identical sizes, generating a bool mask to filter the elements. It was originally used to group identical structural openings, located at a different height.